requirements.txt
fastapi
uvicorn[standard]
rpi.gpio
